final stage changes

+++ betplayer/superagent/AgentLedger.aspx.cs
@@ -197,7 +197,7 @@ namespace betplayer.Super_Agent
                     LedgerTableOrdered.Rows[0]["Date"] = date.Date.ToString().Substring(0, 10);
                     if (l > 0)
                     {
-                        LedgerTableOrdered.Rows[l]["Balance"] = Convert.ToInt32(LedgerTableOrdered.Rows[l - 1]["Balance"]) - Convert.ToInt32(LedgerTableOrdered.Rows[l]["Dabit"]) + Convert.ToInt32(LedgerTableOrdered.Rows[l]["Credit"]);
+                        LedgerTableOrdered.Rows[l]["Balance"] = Convert.ToDecimal(LedgerTableOrdered.Rows[l - 1]["Balance"]) - Convert.ToDecimal(LedgerTableOrdered.Rows[l]["Dabit"]) + Convert.ToDecimal(LedgerTableOrdered.Rows[l]["Credit"]);
 
 
                     }
